LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event-respond-with-argument.https-expected.txt

CONSOLE MESSAGE: line 52: TypeError: window.parent.notify_done is not a function. (In 'window.parent.notify_done('FAIL: no controller')', 'window.parent.notify_done' is undefined)


Harness Error (TIMEOUT), message = null

TIMEOUT respondWith() takes either a Response or a promise that resolves with a Response. Other values should raise a network error. Test timed out
